add switches, rename/moving files
