+++ src/client/static/hydration.tsx
@@ -5,7 +5,8 @@ import { StaticRouter } from 'react-router-dom';
 import { hydrate } from 'react-dom';
 import { RoutingConfig } from '@/client/routes';
 import { Main } from '@/client/main';
-import { tests } from '@/shared/model/experiments';
+import { tests } from '@/shared/model/experiments/abTests';
+import { switches } from '@/shared/model/experiments/abSwitches';
 
 export const hydrateApp = () => {
   const routingConfig: RoutingConfig = JSON.parse(
@@ -19,9 +20,7 @@ export const hydrateApp = () => {
   hydrate(
     <ABProvider
       arrayOfTestObjects={tests}
-      abTestSwitches={{
-        ...{ abAbTestTest: true },
-      }}
+      abTestSwitches={switches}
       pageIsSensitive={false}
       mvtMaxValue={1000000}
       mvtId={mvtId}
