+++ src/server/lib/renderer.tsx
@@ -12,7 +12,8 @@ import { getDefaultServerState, ServerState } from '@/server/models/Express';
 import { FieldError } from '@/server/routes/changePassword';
 import { CsrfErrors } from '@/shared/model/Errors';
 import { ABProvider } from '@guardian/ab-react';
-import { tests } from '@/shared/model/experiments';
+import { tests } from '@/shared/model/experiments/abTests';
+import { switches } from '@/shared/model/experiments/abSwitches';
 
 const assets = getAssets();
 
@@ -85,9 +86,7 @@ export const renderer: (url: string, opts: RendererOpts) => string = (
   const react = ReactDOMServer.renderToString(
     <ABProvider
       arrayOfTestObjects={tests}
-      abTestSwitches={{
-        ...{ abAbTestTest: true },
-      }}
+      abTestSwitches={switches}
       pageIsSensitive={false}
       mvtMaxValue={1000000}
       mvtId={mvtId}
