+++ src/shared/model/experiments/abSwitches.ts
@@ -1,0 +1,3 @@
+export const switches = {
+  abTestTest: true,
+};
