+++ src/shared/model/experiments/abTests.ts
@@ -1,4 +1,4 @@
 import { ABTest } from '@guardian/ab-core';
-import { abTestTest } from './ab-test-test';
+import { abTestTest } from './tests/ab-test-test';
 
 export const tests: ABTest[] = [abTestTest];
